mcp_server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
azure-identity>=1.15.0
azure-mgmt-datafactory>=4.0.0
azure-keyvault-secrets>=4.7.0
azure-mgmt-resource>=23.0.1
python-dotenv==1.0.0
